# api/swagger.yaml
definitions:
  model.Component:
    properties:
      created_at:
        type: integer
      id:
        type: integer
      name:
        type: string
      network_id:
        type: integer
      public_url:
        type: string
      status:
        type: boolean
      type:
        type: string
      updated_at:
        type: integer
    required:
    - network_id
    - type
    type: object
  model.Network:
    properties:
      contract_address_url:
        type: string
      created_at:
        type: integer
      id:
        type: integer
      state_dump_url:
        type: string
      status:
        type: boolean
      updated_at:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Titond-backend application
  title: Titond
  version: "1.0"
paths:
  /api/components:
    get:
      description: Get Component By Type
      operationId: get-component-by-type
      parameters:
      - description: Component type (e.g., l2geth)
        in: query
        name: type
        required: true
        type: string
      - description: Network ID
        in: query
        name: network_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: GetComponentByType
    post:
      consumes:
      - application/json
      description: Create a new component
      operationId: create-component
      parameters:
      - description: Component data to create
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Component'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: CreateComponent
  /api/components/{id}:
    delete:
      description: Delete Component By Id
      operationId: delete-component-by-id
      parameters:
      - description: Component ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: DeleteComponentById
    get:
      description: Get Component By ID
      operationId: get-component-by-id
      parameters:
      - description: Component ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: GetComponentByID
  /api/networks:
    get:
      description: Get networks by page
      operationId: get-networks-by-page
      parameters:
      - description: Network ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: GetNetworksByPage
    post:
      description: Create a new network
      operationId: create-network
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Network'
        "500":
          description: Internal Server Error
      summary: CreateNetwork
  /api/networks/{id}:
    delete:
      description: Delete a network by id
      operationId: delete-network
      parameters:
      - description: Network ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: DeleteNetwork
    get:
      description: Get a network by id
      operationId: get-network-by-id
      parameters:
      - description: Network ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: GetNetworkById
swagger: "2.0"